+++ lib/Model/PaymentTransactionsProductModelSubBasket.php
@@ -1,0 +1,429 @@
+<?php
+
+namespace Secuconnect\Client\Model;
+
+use \ArrayAccess;
+
+/**
+ * PaymentTransactionsProductModelSubBasket
+ *
+ * @category Class
+ * @description Transaction Basket Items
+ * @package  Secuconnect\Client
+ * @author   Swagger Codegen team
+ * @link     https://github.com/swagger-api/swagger-codegen
+ */
+class PaymentTransactionsProductModelSubBasket implements ArrayAccess
+{
+    const DISCRIMINATOR = null;
+
+    /**
+      * The original name of the model.
+      * @var string
+      */
+    protected static $swaggerModelName = 'PaymentTransactionsProductModelSubBasket';
+
+    /**
+      * Array of property to type mappings. Used for (de)serialization
+      * @var string[]
+      */
+    protected static $swaggerTypes = [
+        'tax' => 'int',
+        'quantity' => 'int',
+        'price' => 'int',
+        'reference_id' => 'string',
+        'ean' => 'string',
+        'article_number' => 'string',
+        'item_type' => 'string',
+        'name' => 'string',
+        'total' => 'int'
+    ];
+
+    /**
+      * Array of property to format mappings. Used for (de)serialization
+      * @var string[]
+      */
+    protected static $swaggerFormats = [
+        'tax' => null,
+        'quantity' => null,
+        'price' => null,
+        'reference_id' => null,
+        'ean' => null,
+        'article_number' => null,
+        'item_type' => null,
+        'name' => null,
+        'total' => null
+    ];
+
+    public static function swaggerTypes()
+    {
+        return self::$swaggerTypes;
+    }
+
+    public static function swaggerFormats()
+    {
+        return self::$swaggerFormats;
+    }
+
+    /**
+     * Array of attributes where the key is the local name, and the value is the original name
+     * @var string[]
+     */
+    protected static $attributeMap = [
+        'tax' => 'tax',
+        'quantity' => 'quantity',
+        'price' => 'price',
+        'reference_id' => 'reference_id',
+        'ean' => 'ean',
+        'article_number' => 'article_number',
+        'item_type' => 'item_type',
+        'name' => 'name',
+        'total' => 'total'
+    ];
+
+    /**
+     * Array of attributes to setter functions (for deserialization of responses)
+     * @var string[]
+     */
+    protected static $setters = [
+        'tax' => 'setTax',
+        'quantity' => 'setQuantity',
+        'price' => 'setPrice',
+        'reference_id' => 'setReferenceId',
+        'ean' => 'setEan',
+        'article_number' => 'setArticleNumber',
+        'item_type' => 'setItemType',
+        'name' => 'setName',
+        'total' => 'setTotal'
+    ];
+
+    /**
+     * Array of attributes to getter functions (for serialization of requests)
+     * @var string[]
+     */
+    protected static $getters = [
+        'tax' => 'getTax',
+        'quantity' => 'getQuantity',
+        'price' => 'getPrice',
+        'reference_id' => 'getReferenceId',
+        'ean' => 'getEan',
+        'article_number' => 'getArticleNumber',
+        'item_type' => 'getItemType',
+        'name' => 'getName',
+        'total' => 'getTotal'
+    ];
+
+    public static function attributeMap()
+    {
+        return self::$attributeMap;
+    }
+
+    public static function setters()
+    {
+        return self::$setters;
+    }
+
+    public static function getters()
+    {
+        return self::$getters;
+    }
+
+    /**
+     * Associative array for storing property values
+     * @var mixed[]
+     */
+    protected $container = [];
+
+    /**
+     * Constructor
+     * @param mixed[] $data Associated array of property values initializing the model
+     */
+    public function __construct(array $data = null)
+    {
+        $this->container['tax'] = isset($data['tax']) ? $data['tax'] : null;
+        $this->container['quantity'] = isset($data['quantity']) ? $data['quantity'] : null;
+        $this->container['price'] = isset($data['price']) ? $data['price'] : null;
+        $this->container['reference_id'] = isset($data['reference_id']) ? $data['reference_id'] : null;
+        $this->container['ean'] = isset($data['ean']) ? $data['ean'] : null;
+        $this->container['article_number'] = isset($data['article_number']) ? $data['article_number'] : null;
+        $this->container['item_type'] = isset($data['item_type']) ? $data['item_type'] : null;
+        $this->container['name'] = isset($data['name']) ? $data['name'] : null;
+        $this->container['total'] = isset($data['total']) ? $data['total'] : null;
+    }
+
+    /**
+     * show all the invalid properties with reasons.
+     *
+     * @return array invalid properties with reasons
+     */
+    public function listInvalidProperties()
+    {
+        $invalid_properties = [];
+
+        return $invalid_properties;
+    }
+
+    /**
+     * validate all the properties in the model
+     * return true if all passed
+     *
+     * @return bool True if all properties are valid
+     */
+    public function valid()
+    {
+        return true;
+    }
+
+
+    /**
+     * Gets tax
+     * @return int
+     */
+    public function getTax()
+    {
+        return $this->container['tax'];
+    }
+
+    /**
+     * Sets tax
+     * @param int $tax tax
+     * @return $this
+     */
+    public function setTax($tax)
+    {
+        $this->container['tax'] = $tax;
+
+        return $this;
+    }
+
+    /**
+     * Gets quantity
+     * @return int
+     */
+    public function getQuantity()
+    {
+        return $this->container['quantity'];
+    }
+
+    /**
+     * Sets quantity
+     * @param int $quantity Quantity
+     * @return $this
+     */
+    public function setQuantity($quantity)
+    {
+        $this->container['quantity'] = $quantity;
+
+        return $this;
+    }
+
+    /**
+     * Gets price
+     * @return int
+     */
+    public function getPrice()
+    {
+        return $this->container['price'];
+    }
+
+    /**
+     * Sets price
+     * @param int $price price
+     * @return $this
+     */
+    public function setPrice($price)
+    {
+        $this->container['price'] = $price;
+
+        return $this;
+    }
+
+    /**
+     * Gets reference_id
+     * @return string
+     */
+    public function getReferenceId()
+    {
+        return $this->container['reference_id'];
+    }
+
+    /**
+     * Sets reference_id
+     * @param string $reference_id Unique line identifier, used to update the basket before the transaction is executed
+     * @return $this
+     */
+    public function setReferenceId($reference_id)
+    {
+        $this->container['reference_id'] = $reference_id;
+
+        return $this;
+    }
+
+    /**
+     * Gets ean
+     * @return string
+     */
+    public function getEan()
+    {
+        return $this->container['ean'];
+    }
+
+    /**
+     * Sets ean
+     * @param string $ean ean
+     * @return $this
+     */
+    public function setEan($ean)
+    {
+        $this->container['ean'] = $ean;
+
+        return $this;
+    }
+
+    /**
+     * Gets article_number
+     * @return string
+     */
+    public function getArticleNumber()
+    {
+        return $this->container['article_number'];
+    }
+
+    /**
+     * Sets article_number
+     * @param string $article_number Article number (SKU)
+     * @return $this
+     */
+    public function setArticleNumber($article_number)
+    {
+        $this->container['article_number'] = $article_number;
+
+        return $this;
+    }
+
+    /**
+     * Gets item_type
+     * @return string
+     */
+    public function getItemType()
+    {
+        return $this->container['item_type'];
+    }
+
+    /**
+     * Sets item_type
+     * @param string $item_type Item type:  - `\"article\"` - `\"coupon\"` - `\"sub_transaction\"` - `\"stakeholder_payment\"` - `\"shipping\"` - `\"subscription\"`  _Note: \"coupon\" items have the effect of reducing the amount. They are for coupons managed by the merchant, but not for secucard._
+     * @return $this
+     */
+    public function setItemType($item_type)
+    {
+        $this->container['item_type'] = $item_type;
+
+        return $this;
+    }
+
+    /**
+     * Gets name
+     * @return string
+     */
+    public function getName()
+    {
+        return $this->container['name'];
+    }
+
+    /**
+     * Sets name
+     * @param string $name Descriptive article name
+     * @return $this
+     */
+    public function setName($name)
+    {
+        $this->container['name'] = $name;
+
+        return $this;
+    }
+
+    /**
+     * Gets total
+     * @return int
+     */
+    public function getTotal()
+    {
+        return $this->container['total'];
+    }
+
+    /**
+     * Sets total
+     * @param int $total total
+     * @return $this
+     */
+    public function setTotal($total)
+    {
+        $this->container['total'] = $total;
+
+        return $this;
+    }
+
+    /**
+     * Returns true if offset exists. False otherwise.
+     * @param integer $offset Offset
+     * @return boolean
+     */
+    #[\ReturnTypeWillChange]
+    public function offsetExists($offset)
+    {
+        return isset($this->container[$offset]);
+    }
+
+    /**
+     * Gets offset.
+     * @param integer $offset Offset
+     * @return mixed
+     */
+    #[\ReturnTypeWillChange]
+    public function offsetGet($offset)
+    {
+        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+    }
+
+    /**
+     * Sets value based on offset.
+     * @param integer $offset Offset
+     * @param mixed   $value  Value to be set
+     * @return void
+     */
+    #[\ReturnTypeWillChange]
+    public function offsetSet($offset, $value)
+    {
+        if (is_null($offset)) {
+            $this->container[] = $value;
+        } else {
+            $this->container[$offset] = $value;
+        }
+    }
+
+    /**
+     * Unsets offset.
+     * @param integer $offset Offset
+     * @return void
+     */
+    #[\ReturnTypeWillChange]
+    public function offsetUnset($offset)
+    {
+        unset($this->container[$offset]);
+    }
+
+    /**
+     * Gets the string presentation of the object
+     * @return string
+     */
+    public function __toString()
+    {
+        if (defined('JSON_PRETTY_PRINT')) { // use JSON pretty print
+            return json_encode(\Secuconnect\Client\ObjectSerializer::sanitizeForSerialization($this), JSON_PRETTY_PRINT);
+        }
+
+        return json_encode(\Secuconnect\Client\ObjectSerializer::sanitizeForSerialization($this));
+    }
+}
+